WeatherApiService-1/src/main/java/com/skyapi/weatherforecast/hourly/HourlyWeatherRepository.java
package com.skyapi.weatherforecast.hourly;

import java.util.List;

import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;

import com.skyapi.weatherforecast.common.HourlyWeather;
import com.skyapi.weatherforecast.common.HourlyWeatherId;

public interface HourlyWeatherRepository extends CrudRepository<HourlyWeather, HourlyWeatherId> {
	@Query("""
			SELECT h
			FROM HourlyWeather h
			WHERE h.id.location.code = ?1 AND h.id.hourOfDay > ?2 AND h.id.location.trashed=false
			""")
	public List<HourlyWeather> findByLocationCode(String locationCode, int currentHour);
}
